Extract a RustMatrixClientFactory
